Update to make IE9 work with ranges, and use same fragmentFor function for all IE versions
Unfortunately, IE9 removes comments inside of select tags, so the
select test now fails in IE9. A workaround would be to use text nodes
instead of comments inside of select elements only, but there is not a
clean way of implementing this without requiring developers to specify
whether the container is a select when calling the outerHTML function.